HAPP-1534 - fixed bottom navigation bar icon, color & typography - moved colors to Colors.kt
