+++ app/src/main/java/ca/bc/gov/bchealth/compose/Styles.kt
@@ -3,7 +3,6 @@ package ca.bc.gov.bchealth.compose
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Typography
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
@@ -12,6 +11,10 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import ca.bc.gov.bchealth.R
+import ca.bc.gov.bchealth.compose.theme.darkText
+import ca.bc.gov.bchealth.compose.theme.descriptionGrey
+import ca.bc.gov.bchealth.compose.theme.primaryBlue
+import ca.bc.gov.bchealth.compose.theme.white
 
 // Consider making touch targets at least 48x48dp. https://support.google.com/accessibility/android/answer/7101858?hl=en
 val minButtonSize = 48.dp
@@ -22,20 +25,6 @@ val fonts = FontFamily(
     Font(R.font.bc_sans_italic, weight = FontWeight.Normal, style = FontStyle.Italic),
     Font(R.font.bc_sans_bold_italic, weight = FontWeight.Bold, style = FontStyle.Italic),
 )
-
-val darkText = Color(0xFF313132)
-
-val black = Color(0xFF000000)
-val white = Color(0xFFFFFFFF)
-val blue = Color(0xFF1A5A96)
-val primaryBlue = Color(0xFF003366)
-val lightBlue = Color(0xFFB2C1CF)
-val statusBlue30 = Color(0x4D38598A)
-val descriptionGrey = Color(0xFF6D757D)
-val grey = Color(0xFF606060)
-val greyBg = Color(0xFFF2F2F2)
-val green = Color(0xFF2E8540)
-val red = Color(0xFFD8292F)
 
 val Typography.largeButton: TextStyle
     @Composable
