+++ app/src/main/java/ca/bc/gov/bchealth/compose/theme/Colors.kt
@@ -1,0 +1,21 @@
+package ca.bc.gov.bchealth.compose.theme
+
+import androidx.compose.ui.graphics.Color
+
+val darkText = Color(0xFF313132)
+val black = Color(0xFF000000)
+
+val white = Color(0xFFFFFFFF)
+
+val blue = Color(0xFF1A5A96)
+val primaryBlue = Color(0xFF003366)
+val lightBlue = Color(0xFFB2C1CF)
+val statusBlue30 = Color(0x4D38598A)
+val statusBlue = Color(0xFF38598A)
+
+val descriptionGrey = Color(0xFF6D757D)
+val grey = Color(0xFF606060)
+val greyBg = Color(0xFFF2F2F2)
+
+val green = Color(0xFF2E8540)
+val red = Color(0xFFD8292F)
