+++ app/src/main/java/ca/bc/gov/bchealth/ui/comment/CommentsSummaryUI.kt
@@ -19,9 +19,9 @@ import ca.bc.gov.bchealth.R
 import ca.bc.gov.bchealth.compose.BasePreview
 import ca.bc.gov.bchealth.compose.MyHealthTheme
 import ca.bc.gov.bchealth.compose.MyHealthTypography
-import ca.bc.gov.bchealth.compose.blue
-import ca.bc.gov.bchealth.compose.grey
-import ca.bc.gov.bchealth.compose.greyBg
+import ca.bc.gov.bchealth.compose.theme.blue
+import ca.bc.gov.bchealth.compose.theme.grey
+import ca.bc.gov.bchealth.compose.theme.greyBg
 import ca.bc.gov.common.model.SyncStatus
 import ca.bc.gov.common.utils.toDateTimeString
 import java.time.Instant
