+++ app/src/main/java/ca/bc/gov/bchealth/widget/CommentInputUI.kt
@@ -41,14 +41,14 @@ import ca.bc.gov.bchealth.R
 import ca.bc.gov.bchealth.compose.BasePreview
 import ca.bc.gov.bchealth.compose.MyHealthTheme
 import ca.bc.gov.bchealth.compose.MyHealthTypography
-import ca.bc.gov.bchealth.compose.blue
-import ca.bc.gov.bchealth.compose.darkText
-import ca.bc.gov.bchealth.compose.grey
-import ca.bc.gov.bchealth.compose.greyBg
 import ca.bc.gov.bchealth.compose.minButtonSize
-import ca.bc.gov.bchealth.compose.primaryBlue
-import ca.bc.gov.bchealth.compose.red
-import ca.bc.gov.bchealth.compose.white
+import ca.bc.gov.bchealth.compose.theme.blue
+import ca.bc.gov.bchealth.compose.theme.darkText
+import ca.bc.gov.bchealth.compose.theme.grey
+import ca.bc.gov.bchealth.compose.theme.greyBg
+import ca.bc.gov.bchealth.compose.theme.primaryBlue
+import ca.bc.gov.bchealth.compose.theme.red
+import ca.bc.gov.bchealth.compose.theme.white
 import ca.bc.gov.bchealth.ui.comment.Comment
 import ca.bc.gov.common.model.SyncStatus
 import java.time.Instant
