+++ app/src/main/res/drawable/bottom_nav_selector.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:color="@color/blue" android:state_checked="true"/>
+    <item android:color="@color/status_blue" android:state_checked="true"/>
     <item android:color="@color/grey"/>
 </selector>
