+++ app/src/main/res/drawable/home_selector.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:drawable="@drawable/ic_home_un_selected" android:state_checked="false"/>
+    <item android:drawable="@drawable/ic_home" android:state_checked="true"/>
+</selector>
