+++ app/src/main/res/drawable/ic_dependents.xml
@@ -3,8 +3,7 @@
     android:height="24dp"
     android:viewportWidth="24"
     android:viewportHeight="24">
-
-    <path
-        android:fillColor="#CFCFCF"
-        android:pathData="M15.984,12.969C16.859,12.969 17.813,13.094 18.844,13.344C19.875,13.594 20.828,14 21.703,14.563C22.578,15.125 23.016,15.766 23.016,16.484V18.969H17.016V16.484C17.016,15.109 16.359,13.953 15.047,13.016C15.266,12.984 15.578,12.969 15.984,12.969ZM5.156,13.344C6.188,13.094 7.141,12.969 8.016,12.969C8.891,12.969 9.844,13.094 10.875,13.344C11.906,13.594 12.844,14 13.688,14.563C14.563,15.125 15,15.766 15,16.484V18.969H0.984V16.484C0.984,15.766 1.422,15.125 2.297,14.563C3.172,14 4.125,13.594 5.156,13.344ZM10.078,10.109C9.516,10.703 8.828,11 8.016,11C7.203,11 6.5,10.703 5.906,10.109C5.313,9.516 5.016,8.813 5.016,8C5.016,7.188 5.313,6.484 5.906,5.891C6.5,5.297 7.203,5 8.016,5C8.828,5 9.516,5.297 10.078,5.891C10.672,6.484 10.969,7.188 10.969,8C10.969,8.813 10.672,9.516 10.078,10.109ZM18.094,10.109C17.5,10.703 16.797,11 15.984,11C15.172,11 14.469,10.703 13.875,10.109C13.281,9.516 12.984,8.813 12.984,8C12.984,7.188 13.281,6.484 13.875,5.891C14.469,5.297 15.172,5 15.984,5C16.797,5 17.5,5.297 18.094,5.891C18.688,6.484 18.984,7.188 18.984,8C18.984,8.813 18.688,9.516 18.094,10.109Z" />
+  <path
+      android:pathData="M17.922,7.109C17.516,6.703 17.047,6.5 16.516,6.5C15.984,6.5 15.516,6.703 15.109,7.109C14.703,7.484 14.5,7.938 14.5,8.469C14.5,9 14.703,9.469 15.109,9.875C15.516,10.281 15.984,10.484 16.516,10.484C17.047,10.484 17.516,10.281 17.922,9.875C18.328,9.469 18.531,9 18.531,8.469C18.531,7.938 18.328,7.484 17.922,7.109ZM19,10.953C18.313,11.641 17.484,11.984 16.516,11.984C15.547,11.984 14.719,11.641 14.031,10.953C13.344,10.266 13,9.438 13,8.469C13,7.5 13.344,6.688 14.031,6.031C14.719,5.344 15.547,5 16.516,5C17.484,5 18.313,5.344 19,6.031C19.688,6.688 20.031,7.5 20.031,8.469C20.031,9.438 19.688,10.266 19,10.953ZM8.922,7.109C8.516,6.703 8.047,6.5 7.516,6.5C6.984,6.5 6.516,6.703 6.109,7.109C5.703,7.484 5.5,7.938 5.5,8.469C5.5,9 5.703,9.469 6.109,9.875C6.516,10.281 6.984,10.484 7.516,10.484C8.047,10.484 8.516,10.281 8.922,9.875C9.328,9.469 9.531,9 9.531,8.469C9.531,7.938 9.328,7.484 8.922,7.109ZM10,10.953C9.313,11.641 8.484,11.984 7.516,11.984C6.547,11.984 5.719,11.641 5.031,10.953C4.344,10.266 4,9.438 4,8.469C4,7.5 4.344,6.688 5.031,6.031C5.719,5.344 6.547,5 7.516,5C8.484,5 9.313,5.344 10,6.031C10.688,6.688 11.031,7.5 11.031,8.469C11.031,9.438 10.688,10.266 10,10.953ZM21.531,17.469V16.25C21.531,15.938 20.969,15.563 19.844,15.125C18.75,14.688 17.641,14.469 16.516,14.469C15.703,14.469 14.703,14.656 13.516,15.031C13.859,15.438 14.031,15.844 14.031,16.25V17.469H21.531ZM12.531,17.469V16.25C12.531,15.938 11.969,15.563 10.844,15.125C9.75,14.688 8.641,14.469 7.516,14.469C6.391,14.469 5.266,14.688 4.141,15.125C3.047,15.563 2.5,15.938 2.5,16.25V17.469H12.531ZM16.516,12.969C17.859,12.969 19.281,13.281 20.781,13.906C22.281,14.5 23.031,15.281 23.031,16.25V18.969H1V16.25C1,15.281 1.75,14.5 3.25,13.906C4.75,13.281 6.172,12.969 7.516,12.969C8.984,12.969 10.484,13.313 12.016,14C13.547,13.313 15.047,12.969 16.516,12.969Z"
+      android:fillColor="#606060"/>
 </vector>
