+++ app/src/main/res/drawable/ic_home.xml
@@ -1,9 +1,9 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="19dp"
-    android:height="17dp"
-    android:viewportWidth="19"
-    android:viewportHeight="17">
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
   <path
-      android:pathData="M7.1979,15.655V10.655H11.1979V15.655C11.1979,16.205 11.6479,16.655 12.1979,16.655H15.1979C15.7479,16.655 16.1979,16.205 16.1979,15.655V8.655H17.8979C18.3579,8.655 18.5779,8.085 18.2279,7.785L9.8679,0.255C9.4879,-0.085 8.9079,-0.085 8.5279,0.255L0.1679,7.785C-0.1721,8.085 0.0379,8.655 0.4979,8.655H2.1978V15.655C2.1978,16.205 2.6479,16.655 3.1978,16.655H6.1979C6.7478,16.655 7.1979,16.205 7.1979,15.655Z"
-      android:fillColor="#1A5A96"/>
+      android:pathData="M10.198,19.655V14.655H14.198V19.655C14.198,20.205 14.648,20.655 15.198,20.655H18.198C18.748,20.655 19.198,20.205 19.198,19.655V12.655H20.898C21.358,12.655 21.578,12.085 21.228,11.785L12.868,4.255C12.488,3.915 11.908,3.915 11.528,4.255L3.168,11.785C2.828,12.085 3.038,12.655 3.498,12.655H5.198V19.655C5.198,20.205 5.648,20.655 6.198,20.655H9.198C9.748,20.655 10.198,20.205 10.198,19.655Z"
+      android:fillColor="#38598A"/>
 </vector>
