+++ app/src/main/res/drawable/ic_home_un_selected.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <group>
+    <clip-path
+        android:pathData="M0,0h24v24h-24z"/>
+    <path
+        android:pathData="M12,5.69L17,10.19V18H15V12H9V18H7V10.19L12,5.69ZM12,3L2,12H5V20H11V14H13V20H19V12H22L12,3Z"
+        android:fillColor="#B7B7B7"/>
+  </group>
+</vector>
