+++ app/src/main/res/drawable/ic_services.xml
@@ -6,7 +6,6 @@
   <path
       android:pathData="M10.639,2.797V4.656H8.792C8.353,4.656 8,5.012 8,5.453V7.047C8,7.488 8.353,7.844 8.792,7.844H10.639V9.703C10.639,10.145 10.992,10.5 11.431,10.5H13.014C13.453,10.5 13.806,10.145 13.806,9.703V7.844H15.653C16.092,7.844 16.444,7.488 16.444,7.047V5.453C16.444,5.012 16.092,4.656 15.653,4.656H13.806V2.797C13.806,2.355 13.453,2 13.014,2H11.431C10.992,2 10.639,2.355 10.639,2.797Z"
       android:strokeWidth="1.5"
-      android:fillColor="#ffffff"
       android:strokeColor="#B7B7B7"/>
   <path
       android:pathData="M3,17.5V21C3,21.552 3.448,22 4,22H13.149C15.187,22 17.141,21.187 18.577,19.74L22.176,16.116C22.655,15.633 22.726,14.879 22.347,14.315C21.874,13.612 20.896,13.474 20.247,14.019L15.5,18H10.75C9.831,18 9.312,16.945 9.873,16.217C10.082,15.945 10.406,15.786 10.75,15.786H14.357C14.988,15.786 15.5,15.274 15.5,14.643C15.5,14.012 14.988,13.5 14.357,13.5H10.307C8.619,13.5 7.037,14.321 6.064,15.7L5.5,16.5H4C3.448,16.5 3,16.948 3,17.5Z"
