+++ app/src/main/res/drawable/ic_services_filled.xml
@@ -3,11 +3,10 @@
     android:height="24dp"
     android:viewportWidth="24"
     android:viewportHeight="24">
-  <group>
-    <clip-path
-        android:pathData="M2,2h21v19h-21z"/>
-    <path
-        android:pathData="M10.167,2.891V4.969H8.125C7.64,4.969 7.25,5.366 7.25,5.859V7.641C7.25,8.134 7.64,8.531 8.125,8.531H10.167V10.609C10.167,11.103 10.557,11.5 11.042,11.5H12.792C13.277,11.5 13.667,11.103 13.667,10.609V8.531H15.708C16.193,8.531 16.583,8.134 16.583,7.641V5.859C16.583,5.366 16.193,4.969 15.708,4.969H13.667V2.891C13.667,2.397 13.277,2 12.792,2H11.042C10.557,2 10.167,2.397 10.167,2.891ZM22.406,16.554C23.055,16.068 23.193,15.14 22.716,14.48C22.238,13.819 21.327,13.678 20.678,14.165L16.313,17.438H11.917C11.596,17.438 11.333,17.17 11.333,16.844C11.333,16.517 11.596,16.25 11.917,16.25H12.5H14.833C15.479,16.25 16,15.719 16,15.063C16,14.406 15.479,13.875 14.833,13.875H12.5H11.917H9.062C8.001,13.875 6.973,14.242 6.145,14.914L4.508,16.25H3.167C2.521,16.25 2,16.781 2,17.438V19.813C2,20.469 2.521,21 3.167,21H9H14.852C15.909,21 16.941,20.655 17.794,20.017L22.409,16.554H22.406ZM9.026,16.25H9.058C9.047,16.25 9.036,16.25 9.026,16.25Z"
-        android:fillColor="#38598A"/>
-  </group>
+    <group>
+        <clip-path android:pathData="M2,2h21v19h-21z" />
+        <path
+            android:fillColor="#38598A"
+            android:pathData="M10.167,2.891V4.969H8.125C7.64,4.969 7.25,5.366 7.25,5.859V7.641C7.25,8.134 7.64,8.531 8.125,8.531H10.167V10.609C10.167,11.103 10.557,11.5 11.042,11.5H12.792C13.277,11.5 13.667,11.103 13.667,10.609V8.531H15.708C16.193,8.531 16.583,8.134 16.583,7.641V5.859C16.583,5.366 16.193,4.969 15.708,4.969H13.667V2.891C13.667,2.397 13.277,2 12.792,2H11.042C10.557,2 10.167,2.397 10.167,2.891ZM22.406,16.554C23.055,16.068 23.193,15.14 22.716,14.48C22.238,13.819 21.327,13.678 20.678,14.165L16.313,17.438H11.917C11.596,17.438 11.333,17.17 11.333,16.844C11.333,16.517 11.596,16.25 11.917,16.25H12.5H14.833C15.479,16.25 16,15.719 16,15.063C16,14.406 15.479,13.875 14.833,13.875H12.5H11.917H9.062C8.001,13.875 6.973,14.242 6.145,14.914L4.508,16.25H3.167C2.521,16.25 2,16.781 2,17.438V19.813C2,20.469 2.521,21 3.167,21H9H14.852C15.909,21 16.941,20.655 17.794,20.017L22.409,16.554H22.406ZM9.026,16.25H9.058C9.047,16.25 9.036,16.25 9.026,16.25Z" />
+    </group>
 </vector>
