+++ app/src/main/res/menu/bottom_nav_services_menu.xml
@@ -4,7 +4,7 @@
     <item
         android:id="@+id/home"
         android:enabled="true"
-        android:icon="@drawable/ic_home"
+        android:icon="@drawable/home_selector"
         android:title="@string/home" />
 
     <item
